first-agent/requirements.txt
crewai>=0.130.0
openai>=1.50.0
weave>=0.50.0
requests>=2.30.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
PyGithub>=2.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0 